grafana-agent: switch to agent-bare.yaml
https://github.com/grafana/agent/pull/749 removed
production/kubernetes/agent.yaml, and switched to
production/kubernetes/agent-bare.yaml, which describes a single
Deployment, instead of the previous Deployment and DaemonSet.

The individual kube-state-metrics and node-exporter scrape_configs are
gone - so we don't need to interpolate them in the config anymore.

We still need to ask for the namespace for node-exporter, as this module
still deploys node-exporter.

The repo doesn't contain the ConfigMap contents anymore - these moved to
https://grafana.com/docs/grafana-cloud/quickstart/agent-k8s/k8s_agent_metrics.

Contrary to the `external_labels` `cluster` label added there, we keep
our more powerful `external_labels` map, which of course can contain a
"cluster" key.